Now CLs are computed each iter + alot of fixes
also fixed the NaN thingy

+++ Ldes.m
@@ -1,6 +1,6 @@
 function L = Ldes(W_TO, W_fuel)
 
-    L = sqrt(W_TO *(W_TO - W_fuel));
+    L = sqrt(double(W_TO *(W_TO - W_fuel)));
 
 end
